+++ Libplanet/Blocks/IBlockExcerpt.cs
@@ -1,6 +1,4 @@
 using System.Numerics;
-using Libplanet.Blockchain;
-using Libplanet.Blockchain.Policies;
 
 namespace Libplanet.Blocks
 {
@@ -8,8 +6,6 @@ namespace Libplanet.Blocks
     /// The very essential metadata extracted from a block.  This purposes to determine
     /// the canonical chain.
     /// </summary>
-    /// <seealso cref="IBlockPolicy{T}.CanonicalChainComparer"/>
-    /// <seealso cref="TotalDifficultyComparer"/>
     public interface IBlockExcerpt
     {
         /// <summary>